+++ src/render/ray.rs
@@ -6,7 +6,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn at(&self, t: f64) -> Point {
+    pub fn at(&self, t: f32) -> Point {
         &self.origin + t * &self.direction
     }
 }
